ref(core): move inlineCode to code folder
