+++ static/app/components/core/code/index.tsx
@@ -1,1 +1,2 @@
+export {InlineCode} from './inlineCode';
 export {CodeBlock} from './codeBlock';
