+++ static/app/components/core/code/inlineCode.mdx
@@ -10,7 +10,7 @@ resources:
     WCAG 1.4.4: https://www.w3.org/TR/WCAG22/#resize-text
 ---
 
-import {InlineCode} from 'sentry/components/core/inlineCode';
+import {InlineCode} from 'sentry/components/core/code/inlineCode';
 import {Text} from 'sentry/components/core/text';
 import {Prose} from 'sentry/components/core/text/prose';
 import * as Storybook from 'sentry/stories';
